ExpressionInfo should be retained with a unique_ptr rather than a MallocPtr. https://bugs.webkit.org/show_bug.cgi?id=294060 rdar://152272167
Reviewed by Yusuke Suzuki.

Right now UnlinkedCodeBlock's retain the ExpressionInfo they need for stack traces via a MallocPtr.
MallocPtr's don't run the destructors of the types they point to before freeing them. For ExpressionInfo's
this is mostly fine except for the HashMap in ExpressionInfo, which ends up being leaked.

This patch chanegs UnlinkedCodeBlock to retain ExpressionInfo using a unique_ptr instead. This is a little
tricky because ExpressionInfo has a trailing array of bytes so we have to placement `new` it. In the future
we should consider making ExpressionInfo a TrailingArray type, although it's a little unclear how to do so.

* Source/WTF/wtf/MallocPtr.h:
    Mark that this class shouldn't be used.

Canonical link: https://commits.webkit.org/295872@main