Crash when using libswiftWebKit APIs in simulators https://bugs.webkit.org/show_bug.cgi?id=293831 rdar://152200884
Reviewed by Alexey Proskuryakov.

In https://commits.webkit.org/285750@main, we replaced
libswiftWebKit.dylib with a symlink to WebKit. That symlink installs to

    $(INSTALL_ROOT)$(SYSTEM_CONTENT_PATH)$(SYSTEM_PREFIX)/usr/lib/swift/libswiftWebKit.dylib

so that, on platforms with Cryptex layout, it ends up in the OS Cryptex
image and dyld's shared cache builder records it as a pointer to WebKit.

On iOS simulator, the platform builds with Cryptex layout but there is
no prebuilt shared cache. So, apps that bind to libswiftWebKit crash at
launch trying to find /usr/lib/swift/libswiftWebKit.dylib on disk.

Fix by adding a symlink to /usr/lib/swift for affected simulator builds.
Because the script phase in WebKit.xcodeproj is getting unwieldy, move
it to its own file.

* Source/WebKit/Scripts/create-legacy-swift-overlay-symlink.sh: Added.
* Source/WebKit/WebKit.xcodeproj/project.pbxproj:

Canonical link: https://commits.webkit.org/295876@main