AX: Avoid unnecessary main-thread hits when webpage is visible and active but hasn't yet been focused into https://bugs.webkit.org/show_bug.cgi?id=294056 rdar://152617235
Reviewed by Joshua Hoffman.

When initially loading a page, or when VoiceOver switches from another app back to web, content, it's possible for the
ActivityStaate::IsFocused to not be synced to the accessibility thread in time for the first accessibilityFocusedUIElement
request. We should be more lenient in this case, and still use a tree to retrieve focus if it's both visible and in the
active window under the assumption that it will soon be considered focused (i.e. because the AT actually puts focus
inside the page, which may not have happened yet).

* Source/WebKit/WebProcess/cocoa/WebProcessCocoa.mm:
(WebKit::WebProcess::accessibilityFocusedUIElement):

Canonical link: https://commits.webkit.org/295886@main