Reland [WebAuthn] Implement batching for checking allowCredentials https://bugs.webkit.org/show_bug.cgi?id=293805 rdar://problem/152317844
Reviewed by Brent Fulgham.

This change implements checking the allowCredentials in batches as supported by
the authenticator during getAssertion. This is accomplished with smaller up=0,
get requests to determine if credentials are present on the authenticator.

Then if a credential is detected as present, it is included in the allowCredentials list
in the real request. If no credentials matched, then we already know the call will not
be able to succeed, so we just include the last batch.

Added layout tests for the new behaviors.

This patch includes improved behavior with security keys that by policy always require
user verification. This fixes an issue with the previous iteration of this patch where
we would not retain the pin information during silent requests, causing users to be
prompted for a pin multiple times.

Canonical link: https://commits.webkit.org/295931@main