[GTK][WPE][Skia] Assure SkiaRecordingResult is destroyed on the main thread https://bugs.webkit.org/show_bug.cgi?id=293681
Reviewed by Carlos Garcia Campos.

SkiaRecordingResult is constructed on the main thread and destructed by the
last painting worker thread that made use of it -- that destroys the SkPicture,
as well as the image to fence map from a different thread than the main thread,
leading to (potential) thread-safety issues, let's avoid that.

Covered by existing tests.

* Source/WebCore/platform/graphics/skia/SkiaRecordingResult.h:

Canonical link: https://commits.webkit.org/295939@main