Have to tap twice to navigate to items in the completion list of maps.google.com (overeager content change observer) https://bugs.webkit.org/show_bug.cgi?id=294134 rdar://152194074
Reviewed by Brent Fulgham and Tim Horton.

This is an overeager content change observer issues.
The tap turns the element a grey color, but
that is not an important change, so we should
ignore content changed observer for this part of the page.
All of the search suggestions end up in this labels element,
so that is the best way to find the content.

* Source/WebCore/page/Quirks.cpp:
(WebCore::Quirks::shouldIgnoreContentObservationForClick const):
(WebCore::handleGoogleQuirks):

Canonical link: https://commits.webkit.org/295951@main