REGRESSION(295868@main): TestWebKitAPI.QuickLook.LoadFromMemoryCache is a constant failure rdar://152666886 https://bugs.webkit.org/show_bug.cgi?id=294078
Reviewed by Chris Dumez.

We do not want to register service worker clients whose origin is not HTTP given we check on network process side for cookie access.
We add a check in Document::updateServiceWorkerClientData to validate this origin.
This is in particular important in case of quick look documents that may have a HTTP response (hence they could in theory be service worker clients),
but actually have a non HTTP origin and should not be considered as service worker clients.

Adding it in Document::updateServiceWorkerClientData instead of DocumentLoader::commitData is slightly safer since we can also register service worker clients on resume.

* Source/WebCore/dom/Document.cpp:
(WebCore::Document::updateServiceWorkerClientData):
* Source/WebCore/page/SecurityOrigin.h:
(WebCore::SecurityOrigin::isHTTPFamily const):

Canonical link: https://commits.webkit.org/296042@main