Object.getOwnPropertyDescriptors should return a consistent order for JSFunctions https://bugs.webkit.org/show_bug.cgi?id=286996 rdar://144590087
Reviewed by Yusuke Suzuki.

Section 10.1.11.1 of the spec[1] specifies that we should return keys of objects
in ascending chronological order of creation. When we get the keys of a function
object, we originally return these properties in the correct order. However, the
JSFunction class utilizes lazy reification as an optimization. During this first
access, we reify all three properties, as weaccessing their property descriptor.
This leads to a different ordering when we call getOwnPropertyDescriptors again.
This patch fixes this issue by checking if a lazily reified property has already
been reified and still exists as an own property. If so, we can directly list it
along with the other lazily reified properties. (`prototype` is non-configurable
and thus is exempt from this check, since it cannot be deleted)

The correct order of these properties should be ['length', 'name', 'prototype'].
This is specified by the semantics of `InstantiateOrdinaryFunctionObject`, under
section 15.2.4 of the spec [2]. We start by calling `OrdinaryFunctionCreate` (3)
and assigning the result to `F`. Within this call (specified in 10.2.3 [3]), the
function object is initialized. In step 22, we perform `SetFunctionLength`, thus
adding the `length` property first. Returning from `OrdinaryFunctionCreate` then
brings us to a call to `SetFunctionName` (15.2.4 step 4), which makes `name` the
second property to be initialized. Finally, we call `MakeConstructor` in step 5.
As specified in 10.2.5 [4], we perform `DefinePropertyOrThrow(F,"prototype",...`
in step 6, thus making `prototype` the third and last property to be initialized
for a function object. Thus the proper order is ['length', 'name', 'prototype'].

[1] https://tc39.es/ecma262/multipage/ordinary-and-exotic-objects-behaviours.html#sec-ordinaryownpropertykeys
[2] https://tc39.es/ecma262/#sec-runtime-semantics-instantiateordinaryfunctionobject
[3] https://tc39.es/ecma262/#sec-ordinaryfunctioncreate
[4] https://tc39.es/ecma262/#sec-makeconstructor

* JSTests/stress/object-property-descriptors-order-should-be-consistent.js: Added.
(test.a):
(test):
* JSTests/test262/expectations.yaml:
* Source/JavaScriptCore/runtime/JSFunction.cpp:
(JSC::JSFunction::getOwnSpecialPropertyNames):

Canonical link: https://commits.webkit.org/296060@main