[JSC] Skip index-type update in `tryCloneArrayFromFast` when `fillMode` is `Empty` https://bugs.webkit.org/show_bug.cgi?id=294247
Reviewed by Yusuke Suzuki.

`tryCloneArrayFromFast` adjusts the result’s index type
if the source array contains holes, on the assumption
that the holes will be filled with `undefined`.

This adjustment is valid only when `fillMode` is `Undefined`,
but the current implementation also performs it when
`fillMode` is `Empty`.

This patch changes to prevents the index-type update when
`fillMode` is `Empty`.

                                        TipOfTree                  Patched

array-prototype-concat-empty            7.7375+-0.1340     ^      5.3688+-0.1247        ^ definitely 1.4412x faster

Canonical link: https://commits.webkit.org/296082@main