Add timestamps for API tests in EWS https://bugs.webkit.org/show_bug.cgi?id=294297 rdar://153015147
Reviewed by Ryan Haddad.

To improve the debugging of failures in API tests within EWS, we would like to print timestamps for each logged line in the logs. This change adds a timestamps flag to the run-api-tests command, which will activate timestamps in the API test logs.

* Tools/CISupport/ews-build/steps.py:
(RunAPITests):
* Tools/CISupport/ews-build/steps_unittest.py:

Canonical link: https://commits.webkit.org/296089@main