Allow focusing into cross-origin iframes that have never had user interaction if currently processing a user interaction rdar://153007151 https://bugs.webkit.org/show_bug.cgi?id=294296
Reviewed by Ryosuke Niwa and Alex Christensen.

When tabbing through a document, you might reach the end of the current frame's focusable elements
and dig into an iframe to find the first focusable one there.

When we do this today, if the iframe is cross-origin and has never been interacted with, we find
the element but refuse to actually focus it.

This makes sense to prevent programatic JS focus into a cross-origin iframe, but does not make sense
when the user is driving the focusing themselves (e.g. pressing tab)

Making this change matches other browsers.

* LayoutTests/http/tests/dom/focus-in-cross-origin-iframe-expected.txt: Added.
* LayoutTests/http/tests/dom/focus-in-cross-origin-iframe.html: Added.
* LayoutTests/http/tests/dom/resources/focus-in-cross-origin-iframe-iframe.html: Added.

* Source/WebCore/dom/Element.cpp:
(WebCore::Element::focus): Add an "is NOT processing user gesture" qualification to the early return for
  focus into a cross-origin iframe.

Canonical link: https://commits.webkit.org/296110@main