REGRESSION(296061@main): [iOS] fast/events/ios/select-all-with-existing-selection.html is consistent crash https://bugs.webkit.org/show_bug.cgi?id=294346 rdar://153123388
Reviewed by Aditya Keerthi.

In 296061@main, we started storing the UIVisualEffect to configure the
backdrop view with in a RetainPtr. However, -effectWithStyle: on
UIBlurEffect is not a +1 method, and as such our `adoptNS()` call on its
return value meant that we were not retaining the blur effect. This lead
to an eventual overrelease when the UIVisualEffectView holding the blur
effect was destroyed.

Fix this crash by properly retaining the blur effect.

* LayoutTests/platform/ios/TestExpectations: Remove test gardening.
* Source/WebKit/UIProcess/PDF/WKPDFPageNumberIndicator.mm:
(-[WKPDFPageNumberIndicator initWithFrame:view:pageCount:]):

Canonical link: https://commits.webkit.org/296118@main