[GTK][WPE] Add API to Damage to iterate and get the non-empty rectangles https://bugs.webkit.org/show_bug.cgi?id=294015
Reviewed by Adrian Perez de Castro.

This patch includes many of the changes added in 294236@main to make
Damage iterable that were reverted in 294557@main, but without making
Damage iterable again, since we need to investigate what caused the
artifacts. In the meantime we add nonEmptyRects() to get a copy of
the rects without the empty rectangles and forEachNonEmptyRect() to
iterate the rects skipping the empty ones.

Canonical link: https://commits.webkit.org/296137@main