[CSS] Don't compile functional pseudo-classes like :is() with contextually invalid pseudo-element https://bugs.webkit.org/show_bug.cgi?id=293230 rdar://151605806
Reviewed by Simon Fraser.

This mostly reverts commit b9e946421ea6185c88ef9a881bbad23d4e8fb114 which is
buggy because it makes compiling :is(div::before) the same as compiling :is(div).
(the ::before fragment is "optimized away" as never matching,
but the div fragment is still being compiled and potentially matched)

Most of those cases (pseudo-element in functional pseudo-classes) are already prevented at parsing time,
but nesting resolution can create them after parsing.

The fix here is that whenever there is a pseudo-element (standalone or not) in :is(),
we fallback to the SelectorChecker.

* LayoutTests/TestExpectations:
* Source/WebCore/css/SelectorChecker.cpp:
(WebCore::SelectorChecker::checkOne const):
* Source/WebCore/cssjit/SelectorCompiler.cpp:
(WebCore::SelectorCompiler::addPseudoClassType):

Canonical link: https://commits.webkit.org/296144@main