REGRESSION(295171@main): [ iOS Debug ] fast/speechrecognition/ios/audio-capture.html is a constant crash https://bugs.webkit.org/show_bug.cgi?id=294274 rdar://152981168
Reviewed by Youenn Fablet.

The test crashes for hitting an assertion that ensures audio capture only happens in GPU process. However, the test
explicitly disables audio capture in GPU process. The test was added when SpeechRecognition API did not work with audio
capture in GPU process, but now it has the support and we can just remove the flag in the test to fix the crash.

Also, the test checks audio session category with `internals.audioSessionCategory()`, but that function only returns
category for audio session in web process. Now that audio capture happens in GPU process, web process might not know
the categroy; so removing the check from test.

* LayoutTests/fast/speechrecognition/ios/audio-capture-expected.txt:
* LayoutTests/fast/speechrecognition/ios/audio-capture.html:
* LayoutTests/platform/ios/TestExpectations:

Canonical link: https://commits.webkit.org/296150@main